+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Made it possible to pass a Dossier for the receiveFile-methods.
 * Implemented the WebFiche-object
+* Implemented LadresExport
 
 # Changelog since 1.0.0